requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
cbor==1.0.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
datasets==4.5.0
dill==0.4.0
filelock==3.21.0
FlagEmbedding==1.3.5
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
ijson==3.4.0.post0
inscriptis==2.7.0
ir_datasets==0.5.11
Jinja2==3.1.6
joblib==1.5.3
lxml==5.4.0
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.6.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
packaging==26.0
pandas==3.0.0
pdf2image==1.17.0
peft==0.18.1
pgvector==0.4.2
pillow==12.1.1
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
psycopg2-binary==2.9.11
pyarrow==23.0.0
Pygments==2.19.2
PyMuPDF==1.27.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
sentencepiece==0.2.1
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.10.0
tqdm==4.67.3
transformers==4.44.2
trec-car-tools==2.6
triton==3.6.0
typer==0.23.0
typer-slim==0.23.0
typing_extensions==4.15.0
unlzw3==0.2.3
urllib3==2.6.3
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wheel==0.46.3
xxhash==3.6.0
yarl==1.22.0
zlib-state==0.1.10
google-genai==0.3.0
